app/requirements.txt
fastapi
uvicorn[standard]
redis[async]
kafka-python
asyncpg
